Fixed issue related to radio button
When we go back to a page to update a radio button value, for example from Yes to No the value that gets persisted in DB is “Yes, No”.  This behavior is sporadic, it looks like the Dome is not getting updated properly. I Added logic to only accept one value for the radio buttons.